requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
flask-socketio==5.5.1
python-dotenv==1.1.1
paho-mqtt==2.1.0
requests==2.32.5
